+++ jspp/internal/base/config.go
@@ -20,7 +20,8 @@ type JSPreprocessorConfig struct {
 	}
 	AppConfig          string
 	CssScopeRenderSide string
-	Modules            []string
+	ModulesSrc         []string
+	ModulesLinks       []string
 	ModulesIgnore      []string
 	Plugins            []string
 	Targets            []Target
@@ -32,7 +33,8 @@ func NewJSPreprocessorConfig() kernel.IAppComponentConfig {
 	return &JSPreprocessorConfig{
 		ComponentConfig:    app.NewComponentConfigStruct(),
 		CssScopeRenderSide: "client",
-		Modules:            []string{},
+		ModulesSrc:         []string{},
+		ModulesLinks:       []string{},
 		ModulesIgnore:      []string{},
 		Plugins:            []string{},
 		Targets:            []Target{},